+++ app/src/ai/config.ts
@@ -1,0 +1,5 @@
+export const API_ENDPOINT = 'us-central1-aiplatform.googleapis.com'
+export const PROJECT_ID = 'stay-home-discord-bot'
+export const MODEL_ID = 'chat-bison'
+export const LOCATION_ID = 'us-central1'
+export const dataFileName = 'predict-request.json'
