+++ app/src/ai/helpers.ts
@@ -1,0 +1,12 @@
+import { besties, instructions } from "./data"
+
+export const IGNORED_CONTENT = `I'm not able to help with that, as I'm only a language model.`
+
+export const ALLOWED_CONTENT_TYPES = ['image/jpeg', 'image/png', 'image/gif']
+
+export const getContext = () => {
+  // TODO: experiment to see whethere the bot still pick up the first bestie on the list to mention or not
+  const shuffledBesties = besties.sort(() => Math.random() - 0.5);
+  const bestiesString = shuffledBesties.map(({ name, gender, username }) => `${name} (gender: ${gender}, username: @${username})`).join(', ');
+  return [...instructions, `Some of the besties are: ${bestiesString}`].join('; ')
+}
