+++ app/src/ai/index.ts
@@ -1,0 +1,3 @@
+import generateContent from './apis/vertextAi'
+
+export { generateContent }
